+++ packages/front-end/test/experiments.test.ts
@@ -5,13 +5,7 @@ import {
   percentToDecimal,
   rebalance,
 } from "@/services/utils";
-import {
-  IndexedPValue,
-  adjustPValuesBenjaminiHochberg,
-  adjustPValuesHolmBonferroni,
-  adjustedCI,
-  setAdjustedPValuesOnResults,
-} from "@/services/experiments";
+
 
 describe("variation weighting functions", () => {
   it("getEqualWeights with default precision", () => {
@@ -122,149 +116,5 @@ describe("variation weighting functions", () => {
     expect(floatRound(0.546859483, 3)).toEqual(0.547);
     expect(floatRound(0.546859483, 4)).toEqual(0.5469);
     expect(floatRound(0.546859483, 5)).toEqual(0.54686);
-  });
-});
-
-function mockIndexedPvalue(
-  pvalues: number[],
-  index?: number[]
-): IndexedPValue[] {
-  return pvalues.map((p, i) => {
-    return { pValue: p, index: [index ? index[i] : i] };
-  });
-}
-
-describe("pvalue correction method", () => {
-  it("does HB procedure correctly", () => {
-    expect(
-      adjustPValuesHolmBonferroni(
-        mockIndexedPvalue([0.01, 0.04, 0.03, 0.005, 0.55, 0.6])
-      )
-    ).toEqual(
-      mockIndexedPvalue([0.03, 0.05, 0.12, 0.12, 1, 1], [3, 0, 2, 1, 4, 5])
-    );
-  });
-  it("does BH procedure correctly", () => {
-    expect(
-      adjustPValuesBenjaminiHochberg(
-        mockIndexedPvalue([0.898, 0.138, 0.007, 0.964, 0.538, 0.006, 0.138])
-      ).map((x) => {
-        return { pValue: +x.pValue.toFixed(8), index: x.index };
-      })
-    ).toEqual(
-      mockIndexedPvalue(
-        [0.964, 0.964, 0.7532, 0.2415, 0.2415, 0.0245, 0.0245],
-        [3, 0, 4, 1, 6, 2, 5]
-      )
-    );
-  });
-});
-
-describe("results edited in place", () => {
-  it("pvals and CIs adjusted in place", () => {
-    const results = [
-      {
-        name: "res1",
-        srm: 0.5,
-        variations: [
-          {
-            users: 100,
-            metrics: {
-              met1: { value: 0, cr: 0, users: 0, pValue: 0.025 },
-              met2: { value: 0, cr: 0, users: 0, pValue: 0.03 },
-            },
-          },
-        ],
-      },
-    ];
-    const expectedResultsHB = [
-      {
-        name: "res1",
-        srm: 0.5,
-        variations: [
-          {
-            users: 100,
-            metrics: {
-              met1: {
-                value: 0,
-                cr: 0,
-                users: 0,
-                pValue: 0.025,
-                pValueAdjusted: 0.05,
-              },
-              met2: {
-                value: 0,
-                cr: 0,
-                users: 0,
-                pValue: 0.03,
-                pValueAdjusted: 0.05,
-              },
-            },
-          },
-        ],
-      },
-    ];
-    const expectedResultsBH = [
-      {
-        name: "res1",
-        srm: 0.5,
-        variations: [
-          {
-            users: 100,
-            metrics: {
-              met1: {
-                value: 0,
-                cr: 0,
-                users: 0,
-                pValue: 0.025,
-                pValueAdjusted: 0.03,
-              },
-              met2: {
-                value: 0,
-                cr: 0,
-                users: 0,
-                pValue: 0.03,
-                pValueAdjusted: 0.03,
-              },
-            },
-          },
-        ],
-      },
-    ];
-
-    setAdjustedPValuesOnResults(results, ["met1", "met2"], "holm-bonferroni");
-    expect(results).toEqual(expectedResultsHB);
-    setAdjustedPValuesOnResults(
-      results,
-      ["met1", "met2"],
-      "benjamini-hochberg"
-    );
-    expect(results).toEqual(expectedResultsBH);
-  });
-
-  it("does BH procedure correctly", () => {
-    expect(
-      adjustPValuesBenjaminiHochberg(
-        mockIndexedPvalue([0.898, 0.138, 0.007, 0.964, 0.538, 0.006, 0.138])
-      ).map((x) => {
-        return { pValue: +x.pValue.toFixed(8), index: x.index };
-      })
-    ).toEqual(
-      mockIndexedPvalue(
-        [0.964, 0.964, 0.7532, 0.2415, 0.2415, 0.0245, 0.0245],
-        [3, 0, 4, 1, 6, 2, 5]
-      )
-    );
-  });
-
-  it("adjusts CIs as we expect", () => {
-    const adjCIs95pct = adjustedCI(0.049999999, 0.1, 0.05);
-    expect(adjCIs95pct[0]).toBeGreaterThan(0);
-    expect(adjCIs95pct[1]).toBeLessThan(0.2);
-    expect(adjCIs95pct.map((x) => +x.toFixed(8))).toEqual([0, 0.2]);
-
-    expect(
-      adjustedCI(0.0099999999, 0.1, 0.01).map((x) => +x.toFixed(8))
-    ).toEqual([0, 0.2]);
   });
 });
